[fix] #54 dataSource 클린 아키텍쳐 적용
