+++ app/src/main/java/org/android/bbangzip/data/datasource/remote/BadgesNameRemoteDataSource.kt
@@ -6,9 +6,9 @@ import org.android.bbangzip.data.util.base.BaseResponse
 import javax.inject.Inject
 
 class BadgesNameRemoteDataSource
-    @Inject
-    constructor(
-        private val myPageService: MyPageService,
-    ) {
-        suspend fun getBadgeDetail(badgeName: String): BaseResponse<ResponseBadgeNameDto?> = myPageService.getBadgeDetail(badgeName = badgeName)
-    }
+@Inject
+constructor(
+    private val myPageService: MyPageService,
+) {
+    suspend fun getBadgeDetail(badgeName: String): BaseResponse<ResponseBadgeNameDto?> = myPageService.getBadgeDetail(badgeName = badgeName)
+}
